web/requirements.txt
fastapi==0.103.0
uvicorn==0.27.0
starlette==0.27.0
pydantic==2.4.2
redis==4.5.5
passlib==1.7.4
PyJWT==2.8.0
